Naive Lobby Implementation
Naive server searching via client side and client server interaction that includes a lobby system (lobby.py) where an individual can choose to either host or search for a server.

Baseline.py is a baseline as to how we will implement RTT, and give us a good idea for RTT thresholds for proximity.